+++ Sources/Vapor/Date/Date+RFC1123.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension NSDate {
+    public var rfc1123: String {
+        return RFC1123.shared.formatter.string(from: self)
+    }
+}
