+++ Sources/Vapor/Date/RFC1123.swift
@@ -1,6 +1,7 @@
 import Foundation
 import libc
 
+
 struct RFC1123 {
     static func now() -> String { return NSDate().rfc1123 }
 
@@ -12,11 +13,5 @@ struct RFC1123 {
         formatter.locale = NSLocale(localeIdentifier: "en_US")
         formatter.timeZone = NSTimeZone(abbreviation: "GMT")
         formatter.dateFormat = "EEE',' dd MMM yyyy HH':'mm':'ss 'GMT'"
-    }
-}
-
-extension NSDate {
-    public var rfc1123: String {
-        return RFC1123.shared.formatter.string(from: self)
     }
 }
