+++ Sources/Vapor/Middleware/Middleware.swift
@@ -1,0 +1,3 @@
+public protocol Middleware {
+    func respond(to request: Request, chainingTo next: Responder) throws -> Response
+}
